Extract finish tracking into a Finish_Tracking trait
Move the deferred finish-check mechanism (schedule_finish_check,
run_finish_check, has_unfinished_actions, mark_finish_ready,
get_finish_check_hook) out of the already-large Sync class into a focused
Finish_Tracking trait, matching the existing Sync_Data and Chunker trait
decomposition. A trait keeps the logic in one readable file while still
running in the host's scope, so it reaches the protected members it needs
(should_trigger_finish, get_sync_group_name, add_once, $action_id) with no
coupling or visibility changes and no API break. should_trigger_finish stays
on Sync as the per-import-type policy/extension point.

No behaviour change. e2e 28/271, unit 22/44, PHPStan + PHPCS clean.
